requirements.txt
tensorflow==2.11.*
tensorflow_probability==0.19.0
ruamel.yaml==0.16.*
cython<3
gym==0.26.2
mujoco_py==2.1.2.14
dm_control==1.0.16
wandb
numpy
PyOpenGL
xmltodict